Move IamComputeAccessor to own module
